perf: unify tests into single kernel boot (4 boots -> 2)
Combine basic_boot, heap_allocation, and file_system tests into a
single  binary that boots the kernel once and runs all 10 test
functions with shared heap/framebuffer initialization.

- kernel/tests/common/: pure test function modules (no entry_point!)
- kernel/tests/all.rs: unified entry point with buddy+TLSF setup
- Removes old standalone basic_boot, heap_allocation, file_system files
- should_panic and stack_overflow remain standalone (require panic)
- Fix init_heap to map +1 page for TLSF sentinel block

Before: 4 QEMU boots, 93s total
After:  2 QEMU boots, 49s total (47% faster)

run_tests.sh updated to build all + should_panic.
