+++ kernel/src/allocator.rs
@@ -65,7 +65,8 @@ pub fn init_heap(
 ) -> Result<(), MapToError<Size4KiB>> {
     let page_range = {
         let heap_start = VirtAddr::new(HEAP_START as u64);
-        let heap_end = heap_start + u64::try_from(HEAP_SIZE).unwrap() - 1u64;
+        // Include one extra page for the TLSF sentinel block at heap_end
+        let heap_end = heap_start + u64::try_from(HEAP_SIZE).unwrap();
         let heap_start_page = Page::containing_address(heap_start);
         let heap_end_page = Page::containing_address(heap_end);
         Page::range_inclusive(heap_start_page, heap_end_page)
