--- kernel/tests/basic_boot.rs
@@ -1,34 +1,0 @@
-#![no_std]
-#![no_main]
-#![feature(custom_test_frameworks)]
-#![test_runner(yonti_os::test_runner)]
-#![reexport_test_harness_main = "test_main"]
-
-use bootloader_api::{entry_point, BootInfo};
-use core::panic::PanicInfo;
-use yonti_os::framebuffer;
-
-entry_point!(test_kernel_main, config = &yonti_os::BOOTLOADER_CONFIG);
-
-fn test_kernel_main(boot_info: &'static mut BootInfo) -> ! {
-    yonti_os::init();
-    if let Some(fb) = boot_info.framebuffer.take() {
-        let info = fb.info();
-        let buffer = fb.into_buffer();
-        framebuffer::init(buffer, info);
-    }
-    test_main();
-    yonti_os::hlt_loop();
-}
-
-#[panic_handler]
-fn panic(info: &PanicInfo) -> ! {
-    yonti_os::test_panic_handler(info)
-}
-
-use yonti_os::println;
-
-#[test_case]
-fn test_println() {
-    println!("test_println output");
-}
