+++ kernel/tests/common/basic_boot.rs
@@ -1,0 +1,6 @@
+use yonti_os::println;
+
+#[test_case]
+fn test_println() {
+    println!("test_println output");
+}
